models/onnx_models/requirements.txt
# Веб-развертывание системы детекции повреждений
onnxruntime>=1.15.0
opencv-python>=4.6.0
pillow>=9.0.0
numpy>=1.21.0
torchvision>=0.12.0
